Move all the code to utils, in a designated file
Gets rid of the #[cfg(feature = ...)], centralizes all this related code in an accordingly named file, and utils fits better as a place for this code I think (maybe. Not sure. Serenity is kinda all over the place with its module management)
