+++ src/utils/mod.rs
@@ -6,6 +6,7 @@ mod argument_convert;
 #[cfg(feature = "cache")]
 mod content_safe;
 mod custom_message;
+mod formatted_timestamp;
 mod message_builder;
 #[cfg(feature = "collector")]
 mod quick_modal;
@@ -18,6 +19,7 @@ use std::num::NonZeroU16;
 pub use argument_convert::*;
 #[cfg(feature = "cache")]
 pub use content_safe::*;
+pub use formatted_timestamp::*;
 #[cfg(feature = "collector")]
 pub use quick_modal::*;
 use url::Url;
